# docs/swagger.yaml
definitions:
...
  models.Supplier:
    properties:
      company_id:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      phone:
        type: string
    type: object
  models.SupplierUpdate:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
      phone:
        type: string
    type: object
  models.User:
    properties:
      admin:
        type: boolean
      company_id:
        type: string
      cpf:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      password:
        type: string
      phone:
        type: string
    type: object
  models.UserLogin:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
host: sistemaestoques.fly.dev
info:
  contact:
    email: support@swagger.io
    name: Suporte à API
    url: http://www.swagger.io/support
  description: Esta é uma API de exemplo para um sistema de estoque.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Sistema de Estoque API
  version: "1.0"
paths:
...
  /company/delete/supplier/{id}:
    delete:
      consumes:
      - application/json
      description: Deleta um fornecedor no sistema
      parameters:
      - description: ID do fornecedor
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Fornecedor deletado com sucesso
          schema:
            type: string
        "400":
          description: Requisição Inválida
          schema:
            type: string
        "500":
          description: Erro Interno do Servidor
          schema:
            type: string
      summary: Deletar um fornecedor
      tags:
      - Fornecedor
...
  /company/register/supplier:
    post:
      consumes:
      - application/json
      description: Cria um novo fornecedor no sistema
      parameters:
      - description: Detalhes do cadastro do fornecedor
        in: body
        name: requestBody
        required: true
        schema:
          $ref: '#/definitions/models.Supplier'
      produces:
      - application/json
      responses:
        "201":
          description: Fornecedor criado com sucesso
          schema:
            type: string
        "400":
          description: Requisição Inválida
          schema:
            type: string
        "500":
          description: Erro Interno do Servidor
          schema:
            type: string
      summary: Criar um novo fornecedor
      tags:
      - Fornecedor
  /company/register/user:
    post:
      consumes:
      - application/json
      description: Registrar um novo usuário no sistema
      parameters:
      - description: User registration details
        in: body
        name: requestBody
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Registrar novo usuário
      tags:
      - User
...
  /company/remove/user/{userId}:
    delete:
      consumes:
      - application/json
      description: Deletar usuário do sistema baseado no ID do usuário
      parameters:
      - description: User ID to be deleted
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User deleted successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Deletar usuário
      tags:
      - User
...
  /company/suppliers:
    get:
      consumes:
      - application/json
      description: Lista todos os fornecedores de uma empresa
      parameters:
      - description: ID da empresa
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Lista de fornecedores
          schema:
            items:
              $ref: '#/definitions/models.Supplier'
            type: array
        "400":
          description: Requisição Inválida
          schema:
            type: string
        "500":
          description: Erro Interno do Servidor
          schema:
            type: string
      summary: Listar fornecedores
      tags:
      - Fornecedor
...
  /company/update/supplier:
    put:
      consumes:
      - application/json
      description: Atualiza um fornecedor no sistema
      parameters:
      - description: Detalhes da atualização do fornecedor
        in: body
        name: requestBody
        required: true
        schema:
          $ref: '#/definitions/models.SupplierUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: Fornecedor atualizado com sucesso
          schema:
            type: string
        "400":
          description: Requisição Inválida
          schema:
            type: string
        "500":
          description: Erro Interno do Servidor
          schema:
            type: string
      summary: Atualizar um fornecedor
      tags:
      - Fornecedor
  /company/user/login:
    post:
      consumes:
      - application/json
      description: Login do usuário no sistema
      parameters:
      - description: User login details
        in: body
        name: requestBody
        required: true
        schema:
          $ref: '#/definitions/models.UserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: User logged in successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Login do usuário
      tags:
      - User
swagger: "2.0"